Resolve errors and fix bugs

+++ src/main/java/io/NextGenAudio/Playlist/Service/dto/AddTracksRequest.java
@@ -3,20 +3,20 @@ package io.NextGenAudio.Playlist.Service.dto;
 import java.util.List;
 
 public class AddTracksRequest {
-    private List<Integer> fileIds;
+    private List<Long> fileIds;
     private Integer startPosition;
     private Boolean validateTracks = true;
 
     // Constructors
     public AddTracksRequest() {}
 
-    public AddTracksRequest(List<Integer> fileIds) {
+    public AddTracksRequest(List<Long> fileIds) {
         this.fileIds = fileIds;
     }
 
     // Getters and Setters
-    public List<Integer> getFileIds() { return fileIds; }
-    public void setFileIds(List<Integer> fileIds) { this.fileIds = fileIds; }
+    public List<Long> getFileIds() { return fileIds; }
+    public void setFileIds(List<Long> fileIds) { this.fileIds = fileIds; }
 
     public Integer getStartPosition() { return startPosition; }
     public void setStartPosition(Integer startPosition) { this.startPosition = startPosition; }
